+++ pointers_arrays_strings/main.h
@@ -8,7 +8,7 @@ void rev_string(char *s);
 void puts_half(char *str);
 void print_array(int *a, int n);
 char *_strcpy(char *dest, char *src);
-int _atoi(char *s);
+unsigned long _atoi(char *s);
 char *_strcat(char *dest, char *src);
 
 #endif
